Renaming the models the hard way
